+++ src/ipc.c
@@ -879,6 +879,7 @@ static int show_status(FILE *fp)
 	fprintf(fp, "IGMP query interval  : %d sec\n", igmp_query_interval);
 	fprintf(fp, "IGMP querier timeout : %d sec\n", igmp_querier_timeout);
 	fprintf(fp, "RPT Prune entries    : %u of %u\n", rpt_prune_entries, rpt_prune_limit);
+	fprintf(fp, "Local (S,G) entries  : %u of %u\n", local_sg_entries, local_sg_limit);
 	dump_ssm_ranges(fp);
 	dump_reg_acl(fp);
 	fprintf(fp, "SPT Threshold        : %s\n", spt_threshold.mode == SPT_INF ? "Disabled" : "Enabled");